Revision del funcionamiento de todas las funcionalidades. Se corrigen detalles para que quede optimo. Base para integracion de cotizacion

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -248,7 +248,7 @@ class CreateUsersTable extends Migration
             $table->increments('id');
             $table->string('name');
             $table->string('us_last_name');
-            $table->integer('us_id_card')->unique();
+            $table->bigInteger('us_id_card')->unique();
             $table->string('email', 40)->unique();
             $table->bigInteger('us_cell_phone');
             $table->bigInteger('us_phone');
